create base class for method with group
