+++ src/hidimstat/base_variable_importance.py
@@ -2,6 +2,7 @@ import warnings
 
 from sklearn.base import BaseEstimator
 import numpy as np
+import pandas as pd
 
 
 class BaseVariableImportance(BaseEstimator):
@@ -128,4 +129,98 @@ class BaseVariableImportance(BaseEstimator):
         if self.importances_ is None or self.pvalues_ is None:
             raise ValueError(
                 "The importances need to be called before calling this method"
+            )
+
+
+class BaseVariableImportanceGroup(BaseVariableImportance):
+    """
+    BAsic class for managing methods using groups
+
+    Parameters
+    ----------
+
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.n_groups = None
+        self.groups = None
+        self._groups_ids = None
+
+    def fit(self, X, y=None, groups=None):
+        """Base fit method for methods using groups. Identifies the groups.
+
+        Parameters
+        ----------
+        X: array-like of shape (n_samples, n_features)
+            The input samples.
+        y: array-like of shape (n_samples,)
+            Not used, only present for consistency with the sklearn API.
+        groups: dict, optional
+            A dictionary where the keys are the group names and the values are the
+            list of column names corresponding to each group. If None, the groups are
+            identified based on the columns of X.
+        """
+        if groups is None:
+            self.n_groups = X.shape[1]
+            self.groups = {j: [j] for j in range(self.n_groups)}
+            self._groups_ids = np.array(list(self.groups.values()), dtype=int)
+        elif isinstance(groups, dict):
+            self.n_groups = len(groups)
+            self.groups = groups
+            if isinstance(X, pd.DataFrame):
+                self._groups_ids = []
+                for group_key in self.groups.keys():
+                    self._groups_ids.append(
+                        [
+                            i
+                            for i, col in enumerate(X.columns)
+                            if col in self.groups[group_key]
+                        ]
+                    )
+            else:
+                self._groups_ids = [
+                    np.array(ids, dtype=int) for ids in list(self.groups.values())
+                ]
+        else:
+            raise ValueError("groups needs to be a dictionnary")
+
+    def _check_fit(self, X):
+        """
+        Check if the perturbation method has been properly fitted.
+
+        This method verifies that the perturbation method has been fitted by checking
+        if required attributes are set and if the number of features matches
+        the grouped variables.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+            Input data to validate against the fitted model.
+
+        Raises
+        ------
+        ValueError
+            If the method has not been fitted (i.e., if n_groups, groups,
+            or _groups_ids attributes are missing).
+        AssertionError
+            If the number of features in X does not match the total number
+            of features in the grouped variables.
+        """
+        if self.n_groups is None or self.groups is None or self._groups_ids is None:
+            raise ValueError(
+                "The class is not fitted. The fit method must be called"
+                " to set variable groups. If no grouping is needed,"
+                " call fit with groups=None"
+            )
+        count = 0
+        for index_variables in self.groups.values():
+            if type(index_variables[0]) is int:
+                assert np.all(
+                    np.array(index_variables, dtype=int) < X.shape[1]
+                ), "X does not correspond to the fitting data."
+            count += len(index_variables)
+        if X.shape[1] > count:
+            warnings.warn(
+                "The importance will be computed only for features in the groups."
             )
